avionics images moved to avionics folder
