+++ frontend/src/components/Avionics/AvionicsScreen.js
@@ -1,10 +1,12 @@
 import React from 'react'
 import AvionicsCarousel from "./AvionicsCarousel";
+import AvionicsProducts from "./AvionicsProducts";
 
 const AvionicsScreen = () => {
     return (
         <>
            <AvionicsCarousel/>
+           <AvionicsProducts/>
         </>
     )
 };
