+++ frontend/src/components/Avionics/images/avionics/avionics-1.png
