+++ frontend/src/components/Avionics/images/avionics/avionics-2.png
