+++ frontend/src/components/Avionics/images/avionics/avionics-4.png
