+++ frontend/src/components/Avionics/images/avionics/avionics-images-text.png
