+++ frontend/src/components/Avionics/images/avionics/avionics-line.png
