+++ frontend/src/components/Avionics/images/avionics/roel-circle.png
