+++ frontend/src/components/Avionics/images/avionics/team-why.jpg
