+++ frontend/src/components/ProductsCarousel/ProductsCarouselGarminAd.js
@@ -1,9 +1,9 @@
 import React from "react";
 import Fade from 'react-reveal/Fade';
 import {Col, Row} from "react-bootstrap";
-import AvionicsProducts from "../../Avionics/AvionicsProducts";
+import AvionicsProducts from "../Avionics/AvionicsProducts";
 import './GarminHomeFlag.scss'
-import ProductCarousel from "../../Product(s)/ProductsCarousel/ProductCarousel";
+import ProductCarousel from "./ProductCarousel";
 
 
 const ProductsCarouselGarminAd = () => {
