+++ frontend/src/components/ProductsCarousel/productCarousel.css
