+++ frontend/src/components/SignInRegister/registerScreen.css
