+++ frontend/src/components/Wishlist/WishListScreen.css
@@ -153,7 +153,7 @@ a {
 }
 
 .WishListScreenImg{
-    background-image: url("../../images/bg-graphics/jd-roel-flir-end.png");
+    background-image: url("./jd-roel-flir-end.png");
 }
 
 /*.WishListScreen_graphic {*/
